`test` refactor
- `src` is now `loft`
  - All local imports are now from `loft.`
- `src/tests` is now `test`
  - Tests can be run with `unittest`'s test discovery
    (`python -m unittest`)
