+++ loft/web/server.py
@@ -5,8 +5,8 @@ from threading import Thread
 from flask import Flask, render_template
 from werkzeug.serving import run_simple
 
-from config import Config
-from util.file import open_
+from loft.config import Config
+from loft.util.file import open_
 
 
 class Server:
