+++ loft/web/static/src/base.css
