+++ loft/web/templates/base.html
